+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Disallowed access for SemrushBot and Seekport Crawler robots
 - **Data Request Changes**
   - The dataset selection step is now displayed before going to the proof step
+  - Users receive a data request submitted email after submitting their data
+    request
 
 ## 53.0.0 (May 31, 2021)